Revert "Automated testing (#56)"
The goal of this pull request is to introduce automated testing and potentially make local development easier.

The [.github/workflows/test-integration.yaml](.github/workflows/test-integration.yaml) file is understood by Github as [Actions](https://help.github.com/en/actions/automating-your-workflow-with-github-actions/workflow-syntax-for-github-actions) to run every time code is pushed to the repository. It will use `docker-compose` to install/configure mongo/apache and use [puppeteer](https://pptr.dev/) to control a Chrome web browser simulating a user registering an account, logging in, and navigating to the BCO editor. An example of this can be found [here](https://github.com/corburn/bco_editor/commit/7d9f4fabd8761d3e77ad9fc3318b226a388a0be1/checks?check_suite_id=315819148).

The environment can be tested locally on a computer with [docker installed](https://docs.docker.com/install/#supported-platforms) by replacing 'localhost' with 'mongo' in the cgi-bin/conf/config.json and running `docker-compose up` in the same directory as the docker-compose.yaml file. The server logs will be printed to the terminal. With the  the `-d` flag, docker-compose will run them as daemons in the background. After it starts you will be able to access the server from http://localhost:8080/bco_editor. The server code is mounted inside the apache container so changes will be applied immediately.

The initdb.d/user.js file is read by the mongo container when the database is first initialized to create the database user.

The Dockerfile is used to create the apache container. The path to the Dockerfile is implied in the docker-compose.yaml by the `build: ./` property.

The requirements.txt file is used in the Dockerfile to install the python dependencies.

The example*.js files are used at the end of the test-integration.yaml to control the Chrome browser.

If you're interested in these changes I'm happy to answer any questions.

This reverts commit ab7a8b5b3be147f3e97340fe54353891be7cd656.